refacto(httpClient) - Add HttpClientError, NetworkError, ServerError, and ClientError classes to handle errors in the HttpClient module
- Added HttpClientError class to handle generic errors in the HttpClient module
- Added NetworkError class to handle network-related errors in the HttpClient module
- Added ServerError class to handle server-related errors in the HttpClient module
- Added ClientError class to handle client-related errors in the HttpClient module
- Implemented handler method in the ClientError class to handle specific client errors with status code 403 by redirecting to '/auth/login' URL

- Renamed src/contracts/HttpClient.ts to src/contracts/HttpClient/HttpClient.ts
- Added type exports for HttpClient, BaseOptions, BodyOptions, RequestOptions, and RequestMethod in src/contracts/HttpClient/index.ts

- Added new file src/contracts/HttpClient/ErrorsHandler.ts, which contains the implementations for HttpClientError, NetworkError, ServerError, and ClientError classes

- Modified src/
