+++ src/contracts/HttpClient/index.ts
@@ -1,0 +1,3 @@
+export type {HttpClient, BaseOptions, BodyOptions, RequestOptions, RequestMethod} from './HttpClient.ts'
+
+export {ClientError, HttpClientError, ServerError, NetworkError} from './ErrorsHandler.ts'
